Fixed train/utils

+++ yolo5train/utils/utils.py
@@ -531,15 +531,15 @@ def build_targets(p, targets, model):
                 offsets = torch.cat((z, z[j] + off[0], z[k] + off[1], z[l] + off[2], z[m] + off[3]), 0) * g
 
         # Define
-        b, c = t[:, :2].float().T # <-- changed to float() long().T  # image, class
+        b, c = t[:, :2].long().T # image, class
         gxy = t[:, 2:4]  # grid xy
         gwh = t[:, 4:6]  # grid wh
-        gij = (gxy - offsets).float() # <-- changed to float() long()
-        gi, gj = gij.T  # grid xy indices
-
+        gij = (gxy - offsets).long() 
+        gi, gj = gij.T  # grid xy indices, must be long()
+        
         # Append
         indices.append((b, a, gj, gi))  # image, anchor, grid indices
-        tbox.append(torch.cat((gxy - gij, gwh), 1))  # box
+        tbox.append(torch.cat((gxy - gij.float(), gwh), 1))  # box, <-- changed to float() gij
         anch.append(anchors[a])  # anchors
         tcls.append(c)  # class
